[feat] post api 다이얼로그랑 연결
