+++ app/src/main/java/com/example/aboutme/Myprofile/MainProfileFragment.kt
@@ -32,8 +32,6 @@ class MainProfileFragment : Fragment() {
 
 
 
-
-
         return binding.root
     }
